Fixed Feature: Edit Food/Activity
*Problem
When updating values, it automatically deducts from total calories for the day.

*Fix
Made it remove only when it's already updating the total food/activity (before adding the new value, it removes the old value.)

Added Feature: Live Feed

*Live Feed
implemented a 'view' button that opens a second window, that shows the user
a live feed of whatever they add/edit in their tracker.